kepler-maths/requirements.txt
matplotlib==3.8.3
numpy==1.26.4
pyephem==9.99
